plugins/gateway/f5-bigip/F5BigIPBundle.template/ApiClarityAgent/requirements.txt
certifi==2022.9.24
charset-normalizer==2.0.12
idna==3.4
PyYAML==6.0
requests==2.27.1
urllib3==1.26.12